+++ Sources/SwiftlySalesforce/Errors/ResponseError.swift
@@ -1,11 +1,11 @@
 import Foundation
 
-struct ResponseError: Error {
+public struct ResponseError: Error {
     
-    var code: String? = nil
-    var message: String? = nil
-    var fields: [String]? = nil
-    let metadata: HTTPURLResponse
+    public var code: String? = nil
+    public var message: String? = nil
+    public var fields: [String]? = nil
+    public let metadata: HTTPURLResponse
     
     private let na = "N/A" // 'Not Applicable' or 'Not Available'
 }
@@ -19,7 +19,7 @@ extension ResponseError: LocalizedError {
 
 extension ResponseError: CustomDebugStringConvertible {
     
-    var debugDescription: String {
+    public var debugDescription: String {
         let na = "N/A" // 'Not Applicable' or 'Not Available'
         let fieldStr = fields?.joined(separator: ", ") ?? na
         return "Salesforce response error. Code: \(code ?? na). Message: \(message ?? na). Fields: \(fieldStr). HTTP Status Code: \(metadata.statusCode))"