+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Change Log
 
 ## Version 10.0.1 (March 29, 2022)
-Minor update. Added some inline code documentation and updated README.md.
+Minor update. 
+* Added inline code documentation.
+* Updated README.md.
+* Made `ResponseError` public.
 
 ## Version 10.0.0 (March 28, 2022)
 Major 'breaking' release. Highlights: